Resynchronize transpose button when rerendering song
This was removed in commit 23871bdd57679357786326a2837f65eb5bdf6099
because it was misbehaving with Bootstrap 4.  To fix it, we just need
to skip setting the active class when the input is already checked,
since Bootstrap would then remove the active class with toggleClass
when it was expecting to add it.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>